form event for items not working
