+++ src/scripts/events/formEvents.js
@@ -2,9 +2,15 @@ import {
   createOrder,
   updateOrder
 } from '../../api/orderData';
+import {
+  createNewMenuItem,
+  updateMenuItem,
+} from '../../api/menuData';
 import viewOrders from '../components/orderCards';
+import viewOrder from '../components/viewOrderDetails';
+import orderDetail from '../../api/mergedData';
 
-const fromEvt = () => {
+const formEvt = () => {
   document.querySelector('#main-container').addEventListener('submit', (e) => {
     e.preventDefault();
     if (e.target.id.includes('submit-order')) {
@@ -31,6 +37,47 @@ const fromEvt = () => {
 
       updateOrder(orderObject).then(viewOrders);
     }
+    if (e.target.id.includes('submit-item')) {
+      const [, firebaseKey] = e.target.id.split('--');
+      const itemObject = {
+        itemName: document.querySelector('#item-name').value,
+        itemPrice: document.querySelector('#item-price').value,
+        itemDescription: document.querySelector('#item-description').value,
+        orderId: firebaseKey,
+        itemCategory: document.querySelector('#item-category').value
+      };
+      createNewMenuItem(itemObject)
+        .then((itemArray) => {
+          itemArray.forEach((item) => {
+            if (firebaseKey === item.orderId) {
+              orderDetail(item.orderId).then((orderObject) => viewOrder(orderObject));
+            } else {
+              console.warn('else');
+            }
+          });
+        });
+    }
+
+    if (e.target.id.includes('update-item')) {
+      const [, firebaseKey] = e.target.id.split('--');
+      const itemObject = {
+        itemName: document.querySelector('#item-name').value,
+        itemPrice: document.querySelector('#item-price').value,
+        itemDescription: document.querySelector('#item-description').value,
+        itemCategory: document.querySelector('#item-category').value,
+        firebaseKey
+      };
+      updateMenuItem(itemObject)
+        .then((itemArray) => {
+          itemArray.forEach((item) => {
+            if (firebaseKey === item.firebaseKey) {
+              orderDetail(item.orderId).then((orderObject) => viewOrder(orderObject));
+            } else {
+              console.warn('else');
+            }
+          });
+        });
+    }
   });
 };
-export default fromEvt;
+export default formEvt;
